simulation/requirements.txt
cffi
contourpy==1.3.3
cycler==0.12.1
fenics-basix
fenics-dolfinx
fenics-ffcx
fenics-ufl
fonttools==4.60.1
gmsh==4.14.1
imageio==2.37.0
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.7
mpi4py
networkx==3.5
noise==1.2.2
numpy==2.3.4
packaging==25.0
petsc4py
pillow==12.0.0
pycparser
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
slepc4py
tifffile==2025.10.4
wheel==0.45.1
jsonargparse